overall done

+++ Practice-4/coffe-house/src/components/Coffes.jsx
@@ -1,18 +1,40 @@
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2'
 
 const Coffes = ({ coffe }) => {
 
     const { _id, name, chef, price, photo } = coffe
 
     function deleteHandler(id) {
-        console.log(id)
-        fetch(`http://localhost:2000/coffes/${_id}`, {
-            method: 'DELETE',
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:2000/coffes/${id}`, {
+                    method: 'DELETE',
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+                        }
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-            })
+
     }
 
     return (
